Add Battlereeve Of Dusk december reward card. Update Patches. Fix seasons. Fix collection unique card qtd. Fix collection bottom padding. Change qtd type from Long to Int

+++ app/src/main/kotlin/com/ediposouza/teslesgendstracker/data/Card.kt
@@ -231,7 +231,7 @@ data class CardMissing(
 
         val shortName: String,
         val rarity: CardRarity,
-        val qtd: Long
+        val qtd: Int
 
 )
